Fix up naming in std.  And tests.

+++ rhdl-std/src/impl_get_bit.rs
@@ -3,7 +3,7 @@ use rhdl_core::digital_fn::DigitalFn;
 use rhdl_core::kernel::ExternalKernelDef;
 use rhdl_core::kernel::KernelFnKind;
 
-pub fn get_bit<const N: usize>(x: Bits<N>, i: usize) -> bool {
+pub fn get_bit<const N: usize>(x: Bits<N>, i: u128) -> bool {
     (x.0 >> i) & 1 == 1
 }
 
@@ -29,30 +29,30 @@ mod tests {
     #[test]
     fn test_get_bit() {
         let bits = Bits::<128>::mask();
-        assert!(get_bit(bits,0));
-        assert!(get_bit(bits,127));
-        assert!(get_bit(bits,64));
-        assert!(get_bit(bits,32));
-        assert!(get_bit(bits,16));
-        assert!(get_bit(bits,8));
-        assert!(get_bit(bits,4));
-        assert!(get_bit(bits,2));
-        assert!(get_bit(bits,1));
+        assert!(get_bit(bits, 0));
+        assert!(get_bit(bits, 127));
+        assert!(get_bit(bits, 64));
+        assert!(get_bit(bits, 32));
+        assert!(get_bit(bits, 16));
+        assert!(get_bit(bits, 8));
+        assert!(get_bit(bits, 4));
+        assert!(get_bit(bits, 2));
+        assert!(get_bit(bits, 1));
         let bits = Bits::<32>::mask();
-        assert!(get_bit(bits,0));
-        assert!(get_bit(bits,31));
-        assert!(get_bit(bits,16));
-        assert!(get_bit(bits,8));
-        assert!(get_bit(bits,4));
-        assert!(get_bit(bits,2));
-        assert!(get_bit(bits,1));
+        assert!(get_bit(bits, 0));
+        assert!(get_bit(bits, 31));
+        assert!(get_bit(bits, 16));
+        assert!(get_bit(bits, 8));
+        assert!(get_bit(bits, 4));
+        assert!(get_bit(bits, 2));
+        assert!(get_bit(bits, 1));
         let bits = Bits::<1>::mask();
-        assert!(get_bit(bits,0));
+        assert!(get_bit(bits, 0));
         let bits: Bits<5> = 0b11010.into();
-        assert!(get_bit(bits,4));
-        assert!(get_bit(bits,3));
-        assert!(!get_bit(bits,2));
-        assert!(get_bit(bits,1));
-        assert!(!get_bit(bits,0));
+        assert!(get_bit(bits, 4));
+        assert!(get_bit(bits, 3));
+        assert!(!get_bit(bits, 2));
+        assert!(get_bit(bits, 1));
+        assert!(!get_bit(bits, 0));
     }
 }
